+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 450cc8f - 2018-09-16: Pre-Release v9.3.79.dev5 
+- * 3dfb687 - 2018-09-16: Zooqle torrent provider RSS feed English only results 
+- * bd67a69 - 2018-09-16: Pre-Release v9.3.79.dev5 
 - * f39ab1c - 2018-09-16: Refactored Zooqle torrent provider to handling paging results 
 - * 31a9c17 - 2018-09-16: Pre-Release v9.3.79.dev4 
 - * 69668c7 - 2018-09-16: Hardcoded Zooqle torrent provider to search for english torrents